aw1651 boot0 boot from fpga ok

+++ boot0/load_boot1_from_sdmmc/bsp_mmc_for_boot/mmc.c
@@ -782,7 +782,7 @@ int mmc_startup(struct mmc *mmc)
 {
 	int err;
 	u32 mult, freq;
-	u64 cmult, csize, capacity;
+	unsigned long long cmult, csize, capacity;
 	struct mmc_cmd cmd;
 	char ext_csd[512];
 	int timeout = 1000;
